Fix Clang-Tidy reported issues (#215)
Clean up src/ and test/: clang-format, copyright headers, clang-tidy (Phase 3)

Systematic module-by-module pass (Level 0 through Level 5, plus test/)
applying .clang-format, updating copyright headers to the Besprited
format, and fixing the highest-value clang-tidy/clang-analyzer
findings across the whole tree: typedef -> using, NULL -> nullptr,
C-style casts -> static_cast/const_cast/reinterpret_cast, missing
override/[[nodiscard]]/const annotations, trivial ctors/dtors ->
=default, range-based for loops, push_back -> emplace_back, concatenated
nested namespaces, and narrower underlying types for genuine (non-flag)
enums.

Files with embedded raw JS (cfg.cpp, http_request_wasm.cpp, she.cpp)
and Objective-C++ sources are excluded via .clang-format-ignore, since
clang-format corrupts them (mangles `=>` and `===`); these were
hand-formatted instead.

Real bugs found and fixed along the way:
- Null-deref bugs in MoveCel/CopyCel/BackgroundFromLayer (unchecked
  dynamic_cast/cel->image()), ui/menu.cpp's get_base() call sites, and
  ColorPicker::pickColor()'s FromComposition branch.
- Divide-by-zero guard in ui/grid.cpp's expandStrip().
- Copy-paste bug in convolution_matrix_filter.cpp that bypassed a
  zero-guard 3x (used m_matrix->getDiv() instead of the guarded
  delegate.div).
- UTF-8 decoding bug in base/string.h: signed-char sign extension
  corrupted multi-byte codepoint decoding for non-ASCII input.
- base/string.cpp's from_utf8() used assert() (compiled out by
  NDEBUG) instead of the project's ASSERT() macro (gated on _DEBUG) to
  bounds-check output.
- Latent override-skip bug in ui::TipWindow::onInitTheme (called
  Window::onInitTheme instead of PopupWindow::onInitTheme, masked
  today but a footgun for future changes).
- Two integer-overflow bugs in the scripting API (image_script.cpp,
  imageview_script.cpp): multiplication done in `int` before widening
  to size_t/ptrdiff_t, reachable with large images.
- Missing upper-bound validation on nframes/ncels counts in the
  crash-recovery session loader (read_document.cpp) - a corrupted
  backup file could drive an unbounded read loop.
- Silently-swallowed exception in Session::removeDocument now logged,
  matching the sibling removeFromDisk handler.
- Fallthrough bug in cmd_palette_editor.cpp (missing break silently
  overwrote the green channel when editing red).
- Memory leak in App::initialize(): a `new FrameTag` handed to
  DocumentExporter as an unowned raw pointer was never freed when
  --frame-range was used without an existing tag.
- Uninitialized-memory read in tga_format.cpp's palette reader: an
  unhandled palette_entry_size value left image_palette[i]
  uninitialized before it was read back out; now rejected explicitly.
- Marked 6 leaf classes `final` (SDL2Display, ImportSpriteSheetWindow,
  ToolLoopImpl, PreviewToolLoopImpl, RecentFilesListBox,
  RecentFoldersListBox, ZoomEntry) to make their existing
  no-virtual-dispatch-in-ctor/dtor invariant compiler-enforced.
- Resolved the open bitmask/flag-enum policy question: annotated 14
  genuine flag enums (she::Capabilities, ui::MouseButtons,
  doc::LayerFlags, app::CommandFlags, filters::TiledMode, and others)
  with [[clang::flag_enum]], clearing ~444 false-positive
  EnumCastOutOfRange findings tree-wide; added a flags="true" <enum>
  attribute to the pref.xml code generator for app::gen::SymmetryMode.

Flagged for follow-up, deliberately not fixed: doc/file/col_file.cpp's
palette loader doesn't check feof() after fgetc() (truncated files
corrupt rather than error); several doc/ call sites rely on ASSERT()
alone (a no-op in release builds) to guard non-null preconditions;
msk_file.cpp's legacy MSK reader appears to read 6 bytes past a
size-checked buffer but has no test corpus to verify against; one
gif_format.cpp UninitializedObject finding wasn't chased down given
the module's overall finding volume (878).

Also deferred everywhere, consistently, as large-volume/low-value or
requiring individual case-by-case review rather than a blanket fix:
misc-const-correctness, modernize-use-nodiscard, modernize-avoid-c-
style-cast, modernize-use-auto, bugprone-narrowing-conversions, and
cppcoreguidelines-pro-type-static-cast-downcast (an established
tag-checked "poor man's variant" pattern used throughout she/doc/ui/
render/script).

Build and full test suite verified after each batch; final state is a
clean rebuild with all tests passing.

+++ src/filters/convolution_matrix.cpp
@@ -1,5 +1,5 @@
-// Aseprite
-// Copyright (C) 2001-2015  David Capello
+// Aseprite  | Copyright (C) 2001-2015 David Capello
+// Besprited | Copyright (C) 2026      Veritaware
 //
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License version 2 as
@@ -11,17 +11,18 @@
 
 #include "filters/convolution_matrix.h"
 
-namespace filters {
+namespace filters
+{
 
 ConvolutionMatrix::ConvolutionMatrix(int width, int height)
   : m_width(width)
   , m_height(height)
-  , m_cx(width/2)
-  , m_cy(height/2)
+  , m_cx(width / 2)
+  , m_cy(height / 2)
   , m_div(ConvolutionMatrix::Precision)
   , m_bias(0)
   , m_defaultTarget(0)
-  , m_data(width*height, 0)
+  , m_data(static_cast<std::size_t>(width) * height, 0)
 {
 }
